requirements.txt
discord.py
yt-dlp
ffmpeg-python
APScheduler
feedparser
edge-tts
python-dotenv
flask
aiohttp
